+++ src/components/productCard/index.tsx
@@ -19,7 +19,9 @@ const ProductsCard = ({ product, index, getProductsData }: ProductCard) => {
   const toast = useToast();
 
   const handleDelete = async () => {
-    token &&
+
+    if(confirm("Apakah anda yakin ingin menghapus produk ini?")){
+      token &&
       (await deleteProductById(_id, token)
         .then(() => {
           toast({
@@ -38,6 +40,9 @@ const ProductsCard = ({ product, index, getProductsData }: ProductCard) => {
           });
         })
         .finally(() => getProductsData()));
+    } else {
+      return;
+    }
   };
 
   return (
